+++ app/src/main/java/com/ala158/magicpantry/dialogs/PantryAddShoppingListDialog.kt
@@ -20,7 +20,7 @@ class PantryAddShoppingListDialog : DialogFragment(), DialogInterface.OnClickLis
     private var relatedIngredientId = 0L
 
     interface PantryAddShoppingListDialogListener : Parcelable {
-        fun onPantryAddShoppingListDialogClick(unit: String, name: String, id: Long, amount: Int)
+        fun onPantryAddShoppingListDialogClick(unit: String, name: String, id: Long, amount: Double)
     }
 
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
@@ -66,10 +66,10 @@ class PantryAddShoppingListDialog : DialogFragment(), DialogInterface.OnClickLis
     override fun onClick(dialog: DialogInterface?, which: Int) {
         if (which == DialogInterface.BUTTON_POSITIVE) {
             val amountView = dialogView.findViewById<TextInputEditText>(R.id.add_to_shopping_list_amount)
-            var amount = 0
+            var amount = 0.0
 
             if( amountView.text.toString() != "")
-                amount = amountView.text.toString().toInt()
+                amount = amountView.text.toString().toDouble()
 
             if (pantryAddShoppingListDialogListener != null) {
                 pantryAddShoppingListDialogListener!!.onPantryAddShoppingListDialogClick(
